refactor(pci-rancher): rename component file according naming convention ref: TAPC-987
