--- packages/manager/apps/pci-rancher/src/components/GuideLink/index.ts
@@ -1,1 +1,0 @@
-export * from './guide-utils';
