+++ packages/manager/apps/pci-rancher/src/components/Modal/EditNameModal/EditNameModal.component.tsx
@@ -18,7 +18,7 @@ import { useTranslation } from 'react-i18next';
 
 import { RancherService } from '@/api/api.type';
 import { isValidRancherName } from '@/utils/rancher';
-import Modal from './Modal';
+import Modal from '../Modal.component';
 import { useTrackingAction, useTrackingPage } from '@/hooks/useTrackingPage';
 import { TrackingEvent, TrackingPageView } from '@/utils/tracking';
