+++ packages/manager/apps/pci-rancher/src/components/Modal/GenerateAccessModal/GenerateAccesModal.component.tsx
@@ -18,9 +18,9 @@ import { ODS_THEME_COLOR_INTENT } from '@ovhcloud/ods-common-theming';
 import React, { FC } from 'react';
 
 import { useTranslation } from 'react-i18next';
-import { AccessDetail } from '../../hooks/useGenerateAccessDetail';
+import { AccessDetail } from '../../../hooks/useGenerateAccessDetail';
 import { RancherService } from '@/api/api.type';
-import Modal from './Modal';
+import Modal from '../Modal.component';
 import { useTrackingAction, useTrackingPage } from '@/hooks/useTrackingPage';
 import { TrackingEvent, TrackingPageView } from '@/utils/tracking';
