+++ packages/manager/apps/pci-rancher/src/components/Modal/UpdateSoftwareConfirmModal/UpdateSoftwareConfirmModal.test.tsx
@@ -4,7 +4,7 @@ import { versionsMocked } from '@/_mock_/version';
 import updateTranslation from '@/public/translations/pci-rancher/updateSoftware/Messages_fr_FR.json';
 import dashboardTranslation from '@/public/translations/pci-rancher/dashboard/Messages_fr_FR.json';
 import { render, waitFor } from '@/utils/test/test.provider';
-import UpdateSoftwareModal from './UpdateSoftwareConfirmModal';
+import UpdateSoftwareModal from './UpdateSoftwareConfirmModal.component';
 
 const onConfirmUpdated = jest.fn();
 const onClose = jest.fn();
