+++ packages/manager/apps/pci-rancher/src/components/Table/TextCell/TextCell.component.tsx
@@ -4,7 +4,7 @@ import { ODS_THEME_COLOR_INTENT } from '@ovhcloud/ods-common-theming';
 import { OsdsText } from '@ovhcloud/ods-components/react';
 import { Cell } from '@tanstack/react-table';
 import { RancherService } from '@/api/api.type';
-import './Table.scss';
+import '../Table.scss';
 
 interface DisplayCellInterface {
   cell: Cell<RancherService, unknown>;
