+++ packages/manager/apps/pci-rancher/src/hooks/guideLink/useGuideLink.tsx
