+++ packages/manager/apps/pci-rancher/src/pages/create/index.tsx
@@ -3,9 +3,9 @@ import { useQuery } from '@tanstack/react-query';
 import { useNavigate, useParams } from 'react-router-dom';
 import { PageLayout } from '@ovhcloud/manager-components';
 import { getRancherPlan, getReferenceRancherInfo } from '@/api';
-import CreateRancher from '@/components/layout-helpers/CreateRancher/CreateRancher';
+import CreateRancher from '@/components/layout-helpers/CreateRancher/CreateRancher.component';
 import useCreateRancher from '@/hooks/useCreateRancher';
-import Breadcrumb from '../../components/Breadcrumb/Breadcrumb';
+import Breadcrumb from '../../components/Breadcrumb/Breadcrumb.component';
 import { getRanchersUrl } from '@/utils/route';
 import usePciProject from '@/hooks/usePciProject';
 import { PciProjectPlanCode, RancherService } from '@/api/api.type';
