+++ packages/manager/apps/pci-rancher/src/pages/listing/index.tsx
@@ -16,7 +16,7 @@ import { useTranslation } from 'react-i18next';
 import { useHref, useNavigate, useParams } from 'react-router-dom';
 import { Title } from '@ovhcloud/manager-components';
 import { RancherService } from '@/api/api.type';
-import TableContainer from '@/components/Table/TableContainer';
+import TableContainer from '@/components/Table/TableContainer/TableContainer.component';
 import { useTrackingAction } from '@/hooks/useTrackingPage';
 import { getOnboardingUrl } from '@/utils/route';
 import { TrackingEvent, TrackingPageView } from '@/utils/tracking';
